requirements.txt
selenium==3.141.0
webdriver-manager==3.4.2
djangorestframework
pillow
pandas
xlrd
openpyxl